third_party/blink/web_tests/fast/forms/date/date-format-warning-expected.txt

CONSOLE WARNING: The specified value "parsing" does not conform to the required format, "yyyy-MM-dd".
CONSOLE WARNING: The specified value "Invalid attribute value" does not conform to the required format, "yyyy-MM-dd".
CONSOLE WARNING: The specified value "Invalid attribute value" does not conform to the required format, "yyyy-MM-dd".
CONSOLE WARNING: The specified value "Invalid IDL value" does not conform to the required format, "yyyy-MM-dd".
Invisible INPUT element should not show a format warning.


Visible INPUT element should show a format warning. We'll see three warnings.
PASS successfullyParsed is true

TEST COMPLETE
